+++ libopenxds_io/include/openxds.io/IOBuffer.h
@@ -30,10 +30,10 @@ public:
 	         IOBuffer( const openxds::base::String& str );
 	virtual ~IOBuffer();
 
-	virtual unsigned int read( byte* buffer, unsigned int bufferSize ) const
+	virtual long read( byte* buffer, long bufferSize ) const
 		throw (openxds::io::exceptions::IOException*);
 
-	virtual unsigned int write( const byte* buffer, unsigned int count ) const
+	virtual long write( const byte* buffer, long count ) const
 		throw (openxds::io::exceptions::IOException*);
 
 //	virtual void
